:recycle: SetProfile 컴포넌트화
