--- src/pages/set-profile.tsx
@@ -1,222 +1,0 @@
-import styled from '@emotion/styled';
-import axios from 'axios';
-import { NextPage } from 'next';
-import Head from 'next/head';
-import { useRouter } from 'next/router';
-import React, { useState } from 'react';
-import { useForm } from 'react-hook-form';
-import { API_ENDPOINT, BORDER, BUTTON, COLORS } from '../constants';
-
-const SetProfile: NextPage = () => {
-  const router = useRouter();
-  const {
-    register,
-    handleSubmit,
-    getValues,
-    formState: { errors, isValid },
-  } = useForm<{
-    username: string;
-    email: string;
-    password: string;
-    accountname: string;
-    intro: string;
-    image: string;
-  }>({ mode: 'onChange' });
-
-  const [profileImg, setProfileImg] = useState('/default-profile-w.png');
-
-  const onUploadImg = async () => {
-    const imgData = new FormData();
-    imgData.append('image', getValues().image[0]);
-
-    const uploadImg = await axios(`${API_ENDPOINT}/image/uploadfile`, {
-      method: 'POST',
-      headers: {
-        'Content-Type': 'application/json',
-      },
-      data: imgData,
-    });
-    setProfileImg(`${API_ENDPOINT}/${uploadImg.data.filename}`);
-  };
-
-  const signup = handleSubmit(async (data) => {
-    const res = await axios(`${API_ENDPOINT}/user`, {
-      method: 'POST',
-      headers: {
-        'Content-Type': 'application/json',
-      },
-      data: JSON.stringify({
-        user: {
-          username: data.username,
-          email: data.email,
-          password: data.password,
-          accountname: data.accountname,
-          intro: data.intro,
-          image: profileImg,
-        },
-      }),
-    });
-
-    if (res.data.message === '회원가입 성공') {
-      router.push('/login');
-    }
-  });
-  return (
-    <>
-      <Head>
-        <title>프로필 설정ㅣ낑깡팜</title>
-      </Head>
-      <MainSetProfile>
-        <TitleMain>프로필 설정</TitleMain>
-        <TxtSetProfile>나중에 언제든지 변경할 수 있습니다.</TxtSetProfile>
-        <BoxProfileImg>
-          <ImgProfile src={profileImg} alt="사용자 프로필 이미지" />
-          <Label htmlFor="uploadImag" className="btn-profile-img">
-            <span className="sr-only">프로필 사진 업로드</span>
-          </Label>
-          <input
-            type="file"
-            id="uploadImag"
-            accept="image/*"
-            // onChange={onUploadImg}
-            className="sr-only"
-            {...register('image', {
-              onChange: onUploadImg,
-            })}
-          />
-        </BoxProfileImg>
-        <FormSetProfile onSubmit={signup}>
-          <BoxInp>
-            <Label htmlFor="name">
-              사용자 이름
-              <input
-                type="text"
-                id="name"
-                placeholder="2~10자 이내여야 합니다."
-                minLength={2}
-                maxLength={12}
-                required
-              />
-            </Label>
-          </BoxInp>
-          <BoxInp>
-            <Label htmlFor="id">
-              계정 ID
-              <input
-                type="text"
-                id="id"
-                placeholder="영문, 숫자, 특주문자(.),(_)만 사용 가능합니다."
-                required
-              />
-            </Label>
-            <TxtError>
-              *영문, 숫자, 밑줄 및 마침표만 사용할 수 있습니다.
-            </TxtError>
-            <TxtError>*이미 사용 중인 ID 입니다.</TxtError>
-          </BoxInp>
-          <BoxInp>
-            <Label htmlFor="introduce">
-              소개
-              <input
-                type="text"
-                id="introduce"
-                placeholder="자신과 판매할 상품에 대해 소개해 주세요!"
-                required
-              />
-            </Label>
-          </BoxInp>
-          <BtnStart type="submit">낑깡팜 시작하기</BtnStart>
-        </FormSetProfile>
-      </MainSetProfile>
-    </>
-  );
-};
-
-export default SetProfile;
-
-const MainSetProfile = styled.main`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  padding: 30px 34px;
-`;
-const TitleMain = styled.h2`
-  margin-bottom: 40px;
-  color: ${COLORS.black};
-  font-size: 24px;
-  font-weight: 700;
-`;
-const TxtSetProfile = styled.p`
-  font-size: 14px;
-`;
-const BoxProfileImg = styled.div`
-  position: relative;
-  margin: 30px 0;
-`;
-const ImgProfile = styled.img`
-  width: 110px;
-  height: 110px;
-  border-radius: 50%;
-  background-color: ${COLORS.light_gray};
-  object-fit: cover;
-`;
-const FormSetProfile = styled.form`
-  width: 100%;
-`;
-const BoxInp = styled.div`
-  &:not(:first-of-type) {
-    margin-top: 16px;
-  }
-`;
-const Label = styled.label`
-  font-size: 12px;
-
-  &.btn-profile-img {
-    position: absolute;
-    right: 0;
-    bottom: 0;
-    width: 36px;
-    height: 36px;
-    border-radius: 50%;
-    background: url('/icons/img/image.svg') no-repeat 50% 50%
-      ${COLORS.accent_green};
-  }
-
-  & input {
-    display: block;
-    width: 100%;
-    padding: 10px 0 8px;
-    border: 0;
-    border-bottom: ${BORDER.basic};
-    color: ${COLORS.black};
-    font-size: 14px;
-  }
-  & input::placeholder {
-    color: ${COLORS.placeholder};
-  }
-  & input:focus,
-  & input:active {
-    border-color: ${COLORS.accent_green};
-  }
-`;
-const TxtError = styled.span`
-  display: inline-block;
-  margin-top: 6px;
-  color: ${COLORS.error};
-  font-size: 12px;
-`;
-const BtnStart = styled.button`
-  width: 100%;
-  margin: 30px 0 20px;
-  padding: 13px 0;
-  border-radius: 44px;
-  background-color: ${BUTTON.disabled_color};
-  color: ${BUTTON.color};
-  font-weight: 700;
-  transition: all 0.3s;
-
-  &:hover,
-  &:active {
-    background-color: ${BUTTON.background_color};
-  }
-`;
