+++ src/pages/signup.tsx
@@ -2,6 +2,7 @@ import { NextPage } from 'next';
 import { useState } from 'react';
 import Head from 'next/head';
 import { SignupEmail } from '../components/signup/SignupEmail';
+import { SetProfile } from '../components/signup/SetProfile';
 
 const Signup: NextPage = () => {
   const [pages, setPages] = useState(true);
@@ -10,15 +11,12 @@ const Signup: NextPage = () => {
   return (
     <>
       <Head>
-        <title>이메일로 회원가입ㅣ낑깡팜</title>
+        <title>{pages ? '이메일로 회원가입' : '프로필 설정'}ㅣ낑깡팜</title>
       </Head>
       {pages ? (
         <SignupEmail setPages={setPages} setUserInfo={setUserInfo} />
       ) : (
-        <div>
-          <p>프로필 설정</p>
-          {userInfo.email} / {userInfo.password}
-        </div>
+        <SetProfile userInfo={userInfo} />
       )}
     </>
   );
